requirements.txt
Flask==3.0.0
Flask-RESTX==1.3.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.6.0
pymongo[srv]==4.6.0
certifi==2023.11.17
python-dotenv==1.0.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
google-generativeai==0.3.2
google-cloud-speech==2.23.0
ffmpeg-python==0.2.0
python-dateutil==2.8.2
requests==2.31.0
email-validator==2.1.0
werkzeug==3.0.1
gunicorn==21.2.0
